Move pricing database ownership to core
